requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-allauth==0.56.1
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-js-asset==2.1.0
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==8.3.1
psycopg2==2.9.7
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
